Cleanup swig files.
