+++ gocv/gocv_calib3d.i
@@ -1,0 +1,5 @@
+%{
+#include "gocv_calib3d.hpp"
+%}
+
+%include "gocv_calib3d.hpp"
