+++ gocv/gocv_core_test.go
